requirements.txt
python-telegram-bot
schedule
mcstatus
